Add .gitignore file for Python project
This commit adds a .gitignore file with standard rules for Python projects. It ignores various types of unnecessary files such as .pyc files, __pycache__ directories, and Django logs that don't need to be included in the version control. This will result in a cleaner repository by excluding irrelevant files and folders and preventing them from being tracked and included in commits. Furthermore, the .gitignore serves to optimize collaboration by ensuring that each collaborator’s working environment-specific files are not pushed to the shared repository.