Add initial project configuration and job posting functionality
A large change that sets up the initial Django project and its configuration. This includes creation of the jobs app, its models, and view for the portfolio page. The project's settings.py file is also set up to use Postgres as the database. Docker compose file is added to set up a local Postgres database and required Pycharm configurations are added.

The jobs app added includes the ability to post jobs with images and a summary as per the Job model. Functionality to view these postings is also added.

Bootstrap layout for the site and a number of other defaults set up by Django also form part of this commit (e.g. wsgi and asgi files for deployment).