Commit Message:
Configure Django for 'myporfolio' project and update models, templates, and static files

This commit configures Django for the 'myporfolio' project updating the different parts for better functioning. It updates the .iml file to correctly define django facets and template configurations. The models.py is also updated to extend the Job model by adding a description field. The home.html template is significantly refactored, removing unnecessary bootstrap templates and configuring it to properly load static files and display job objects from a database. The settings.py file is updated to correctly define paths of static and media files. Some unused file(color-modes.js) is also removed in this commit. These all changes are useful in improving the overall functionality and layout of the portfolio page as well as the handling of static and media files.